Refactor QuestionEntry and DeleteQnBtn
